refactor: Move DbCandidate utils to fixtures
