+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/utils/BatchScanUtilTest.java
@@ -3,7 +3,7 @@ package no.sikt.nva.nvi.common.utils;
 import static java.util.Collections.emptyList;
 import static no.sikt.nva.nvi.test.TestUtils.createNumberOfCandidatesForYear;
 import static no.sikt.nva.nvi.test.TestUtils.getYearIndexStartMarker;
-import static no.sikt.nva.nvi.test.TestUtils.randomCandidate;
+import static no.sikt.nva.nvi.common.db.DbCandidateFixtures.randomCandidate;
 import static no.sikt.nva.nvi.test.TestUtils.randomIntBetween;
 import static no.sikt.nva.nvi.test.TestUtils.randomYear;
 import static no.sikt.nva.nvi.test.TestUtils.setupReportedCandidate;
