+++ nvi-commons/src/testFixtures/java/no/sikt/nva/nvi/common/db/CandidateDaoFixtures.java
@@ -1,0 +1,3 @@
+package no.sikt.nva.nvi.common.db;
+
+public class CandidateDaoFixtures {}
